Move window system traits to emacs crates
