+++ rust_src/crates/emacs/build.rs.in
@@ -1,6 +1,7 @@
 extern crate bindgen;
 
 use bindgen::Formatter;
+use cfg_aliases::cfg_aliases;
 use std::cmp::max;
 use std::env;
 use std::fs::File;
@@ -338,4 +339,26 @@ fn main() {
     generate_definitions();
     generate_globals();
     generate_bindings();
+
+    // Setup cfg aliases
+    cfg_aliases! {
+        android_platform: { target_os = "android" },
+        wasm_platform: { target_arch = "wasm32" },
+        macos_platform: { target_os = "macos" },
+        ios_platform: { target_os = "ios" },
+        windows_platform: { target_os = "windows" },
+        apple: { any(target_os = "ios", target_os = "macos") },
+        free_unix: { all(unix, not(apple), not(android_platform)) },
+
+        x11_platform: { all(feature = "x11", free_unix, not(wasm), use_winit)},
+        wayland_platform: { all(feature = "wayland", free_unix, not(wasm), use_winit) },
+
+        winit: { any(feature = "window-system-winit") },
+        pgtk: { all(feature = "window-system-pgtk", not(winit)) },
+        window_system: { any(winit, pgtk) },
+
+        surfman: { all(feature = "surfman") },
+        glutin: { all(feature = "glutin") },
+        gtk3: { all(feature = "gtk3") },
+    }
 }
