+++ rust_src/crates/gl-renderer/src/lib.rs
@@ -13,20 +13,6 @@ pub mod frame;
 
 mod cursor;
 
-#[cfg(have_window_system)]
-pub mod window_system {
-
-    #[cfg(window_system_pgtk)]
-    mod pgtk {
-        pub mod frame;
-    }
-
-    #[cfg(window_system_winit)]
-    mod winit {
-        pub mod frame;
-    }
-}
-
 pub mod gl_renderer;
 pub mod image;
 mod image_cache;
@@ -36,26 +22,6 @@ mod fringe;
 mod glyph;
 mod texture;
 mod util;
-
-pub mod gl {
-    pub mod context;
-
-    pub mod context_impl {
-        #[cfg(use_glutin)]
-        pub use crate::gl::context_impl::glutin::*;
-        #[cfg(use_gtk3)]
-        pub use crate::gl::context_impl::gtk3::*;
-        #[cfg(use_surfman)]
-        pub use crate::gl::context_impl::surfman::*;
-
-        #[cfg(use_glutin)]
-        pub mod glutin;
-        #[cfg(use_gtk3)]
-        pub mod gtk3;
-        #[cfg(use_surfman)]
-        pub mod surfman;
-    }
-}
 
 #[cfg(not(test))]
 include!(concat!(env!("OUT_DIR"), "/c_exports.rs"));
