+++ rust_src/crates/winit-term/Cargo.toml
@@ -20,37 +20,12 @@ lazy_static = "1.4"
 app_units = "0.7"
 bit-vec = "0.6.3"
 once_cell = "1.16.0"
-winit = { version = "0.29", default-features = false, features = ["rwh_05"]}
 raw-window-handle.workspace = true
-arboard = { version = "3.3"}
+winit.workspace = true
 
 [build-dependencies]
 ng-bindgen = { path = "../../ng-bindgen" }
-cfg_aliases = "0.2"
-
-[target.'cfg(all(unix, not(target_os = "macos")))'.dependencies]
-wayland-sys = {version = "0.30", features = ["client", "dlopen"]}
-gtk = { workspace = true, optional = true }
-gdk = { workspace = true, optional = true }
-gl_loader = { version = "0.1.2", optional = true }
+cfg_aliases.workspace = true
 
 [target.'cfg(target_os = "macos")'.dependencies]
 core-foundation = "0.9"
-
-[target.'cfg(all(unix, not(target_os = "macos")))'.dependencies.x11]
-features = ["xlib"]
-version = "2.20"
-optional = true
-
-[features]
-default = ["wayland"]
-x11 = [
-  "dep:x11",
-  "winit/x11",
-]
-wayland = [
-  "arboard/wayland-data-control",
-  "winit/wayland",
-  "winit/wayland-dlopen",
-  "winit/wayland-csd-adwaita",
-]
